Next Era AI/requirements.txt
streamlit
langchain
openai
chromadb
python-dotenv
pypdf
pillow
faiss-cpu
serper-dev
PyMuPDF
